Everything GCC and Clang say that MSVC does not, and macOS gets thread times
With -k 0 the two broken jobs named all of it at once.

GCC's -Wextra objects to an unnamed enum's constant sharing a conditional with
a uint16_t, which is how three of the voxel tests spell "this block, or air".
Casts. libstdc++ does not reach <cmath> through the headers test_transmission
already had, so std::sqrt was undeclared; one include. And -Wnull-dereference
reads schroeder_rt60's decay curve as possibly empty, which it is when there is
nothing after `start`: now it says so and returns 0.

macOS has no pthread_getcpuclockid, so the thread registry had never compiled
there. The Mach port answers the same question through thread_info, and
pthread_mach_thread_np hands it over without taking a reference, so `clock`
carries the port on Apple and there is nothing extra to release.

+++ native/tests/core/test_pathing_bake.cpp
@@ -101,13 +101,13 @@ private:
             for (int x = x0 - 1; x <= x1 + 1; ++x) {
                 for (int y = floor - 1; y <= floor + tall; ++y) {
                     const bool shell = x == x0 - 1 || x == x1 + 1 || z == z0 - 1 || z == z1 + 1 || y == floor - 1 || y == floor + tall;
-                    set(x, y, z, shell ? wall : Air);
+                    set(x, y, z, shell ? wall : static_cast<uint16_t>(Air));
                 }
                 for (int y = floor + tall + 1; y < kRegion; ++y) {
                     set(x, y, z, Air);  // nothing above the roof
                 }
                 for (int y = 0; y < floor - 1; ++y) {
-                    set(x, y, z, get(x, y, z) == Air ? Stone : get(x, y, z));  // solid ground under it
+                    set(x, y, z, get(x, y, z) == Air ? static_cast<uint16_t>(Stone) : get(x, y, z));  // solid ground under it
                 }
             }
         }
